move semester-detail into web/modules
