+++ modules/web/modules/semester-detail/semester-detail.scss
