+++ modules/web/routes/student/routes/search/search-sidebar.js
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react'
 import withRouter from 'react-router/lib/withRouter'
-import CourseSearcher from '../../../../../modules/course-searcher/course-searcher-container'
+import CourseSearcher from '../../../../modules/course-searcher/course-searcher-container'
 
 function CourseSearcherSidebar(props) {
 	const { studentId } = props.params
